app/requirements.txt
azure-storage-blob==12.19.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-ai-formrecognizer==3.3.2
pyodbc==5.1.0
python-dotenv==1.0.1
msrest==0.7.1
Pillow==10.3.0
requests==2.31.0
tabulate==0.9.0
